refactor: implement TODO.bugs/09 (Parser split) and TODO.bugs/12 (collection registry)
Completes the last two deferred TODOs from the 2026-07-14 architecture
audit. All 13 TODOs are now implemented.

TODO 09: Split Parser (P2)
- Extract ~250 Parsanol grammar rules from the inner Parser::Parser class
  to Expressir::Express::Grammar::Parser (699 lines, focused on grammar)
- Extract schema-block splitting (extract_schema_blocks, parse_schema_block,
  skip_spaces, skip_ws_and_comments) to Expressir::Express::SchemaBlockScanner
  (132 lines, a self-contained state-machine module)
- Parser.rb shrinks from 1044 to 238 lines — a thin facade that delegates
  to Grammar::Parser and SchemaBlockScanner
- All public methods (from_file, from_exp, etc.) remain backward-compatible
- Updated parser_spec.rb references from Parser::Parser to Grammar::Parser

Architecture: three deep modules behind one facade. Grammar changes land
in grammar/parser.rb; I/O orchestration lands in parser.rb; block-scanning
logic lands in schema_block_scanner.rb. Each file has a single concern.

TODO 12: Generate COLLECTION_REGISTRY from model declarations (P3)
- Add `collection_attributes` class-level macro to ModelElement: each
  model class declares which attributes hold child collections, and the
  declaration auto-registers into ModelElement.collection_registry
- All 11 model classes that previously appeared in the hand-maintained
  COLLECTION_REGISTRY hash now declare their own collections:
  Schema, Entity, Function, Procedure, Rule, Type, ExpFile,
  Statements::Compound, Statements::If, Statements::Alias,
  Statements::Repeat
- NodePositionIndex::COLLECTION_REGISTRY is now a reference to
  ModelElement.collection_registry (same mutable hash object), not a
  duplicated hash. Adding a new model class with collection_attributes
  automatically makes it appear in the registry.
- RemarkAttacher and ScopeResolver reference NodePositionIndex::
  COLLECTION_REGISTRY — transparently updated
- New spec: spec/expressir/model/collection_attributes_spec.rb (4 examples)
  verifying the registry contents, the identity of references, and the
  live-registration behavior

Architecture: the "what collections does this class have" fact now lives
on the model class itself (OCP: new class = new declaration, no central
hash to edit). Single source of truth eliminates the parallel taxonomy
between marker modules and COLLECTION_REGISTRY.

All 1455 examples pass (4 pre-existing grammar pendings). RuboCop clean.
