+++ lib/expressir/model/statements/if.rb
@@ -4,6 +4,7 @@ module Expressir
       # Specified in ISO 10303-11:2004
       # - section 13.7 If ... Then ... Else statement
       class If < ModelElement
+        collection_attributes :statements
         attribute :expression, ModelElement
         attribute :statements, ModelElement, collection: true
         attribute :else_statements, ModelElement, collection: true
